Refactor more of InvocationSummary out.
